Corrected DataMember Name attribute for MoreResultsAvailable

+++ HubSpot.NET/Core/ListRequestOptions.cs
@@ -40,7 +40,7 @@ namespace HubSpot.NET.Core
         /// The return DTO from List contains the current "offset" that you can inject into your next list call 
         /// to continue the listing process
         /// </remarks>
-        public int? Offset { get; set; } = null;
+        public long? Offset { get; set; } = null;
 
         public List<string> PropertiesToInclude { get; set; } = new List<string>();
     }
